Implemented the basics of the ComdirectApi
Currently it's possible to login, refresh a session and end a session.

From now on the task will be to implement the rest of the api.

Also, I should find a mechanism to refresh the session automatically (with a separate thread). But for this there is some kind of synchronisation needed, and I need some time to figure out the best solution for this.